refactor(server): remove dead code and apply minimal DIP for Skill/Issue
Delete confirmed unreachable production symbols, sink same-package test-only
helpers out of production files, mark unwired sandbox protocol stubs, and
move IsPlatformAuthEnvKey into envauth so services no longer depend on runtime
for that rule. Cross-package test factories (NewMemory, OpenSQLiteTest, etc.)
are intentionally retained.
