+++ server/internal/gateshare/ratelimit_helpers_test.go
@@ -1,0 +1,8 @@
+package gateshare
+
+// LenForTest returns the number of tracked IP+bucket keys (tests / diagnostics).
+func (l *IPLimiter) LenForTest() int {
+	l.mu.Lock()
+	defer l.mu.Unlock()
+	return len(l.byKey)
+}
