+++ server/internal/models/token_usage_helpers_test.go
@@ -1,0 +1,18 @@
+package models
+
+// sumTokenUsageByModel returns the component-wise total across buckets.
+// nil byModel → nil; non-nil (incl. empty) → non-nil total (reported).
+// Test-only helper (production aggregation uses EffectiveUsageByModel paths).
+func sumTokenUsageByModel(byModel TokenUsageByModel) *TokenUsage {
+	if byModel == nil {
+		return nil
+	}
+	out := &TokenUsage{}
+	for _, u := range byModel {
+		out.InputTokens += u.InputTokens
+		out.OutputTokens += u.OutputTokens
+		out.CacheReadTokens += u.CacheReadTokens
+		out.CacheWriteTokens += u.CacheWriteTokens
+	}
+	return out
+}
