+++ server/internal/services/channels_config.go
@@ -147,41 +147,11 @@ func (s *ChannelConfigService) ListByProject(projectID string) ([]ChannelConfigD
 	return out, nil
 }
 
-// GetByID returns one channel DTO or ErrChannelNotFound.
-func (s *ChannelConfigService) GetByID(id string) (ChannelConfigDTO, error) {
-	var row models.ChannelConfig
-	if err := s.db.First(&row, "id = ?", strings.TrimSpace(id)).Error; err != nil {
-		if errors.Is(err, gorm.ErrRecordNotFound) {
-			return ChannelConfigDTO{}, ErrChannelNotFound
-		}
-		return ChannelConfigDTO{}, err
-	}
-	return toChannelDTO(row), nil
-}
-
 // GetPrimaryByProject returns the primary channel, or nil when the project has none.
 func (s *ChannelConfigService) GetPrimaryByProject(projectID string) (*ChannelConfigDTO, error) {
 	var r models.ChannelConfig
 	err := s.db.Where("project_id = ? AND is_primary = ?", strings.TrimSpace(projectID), true).
 		Order("created_at asc").First(&r).Error
-	if errors.Is(err, gorm.ErrRecordNotFound) {
-		return nil, nil
-	}
-	if err != nil {
-		return nil, err
-	}
-	dto := toChannelDTO(r)
-	return &dto, nil
-}
-
-// GetByAgent returns the channel bound to agentName, or nil.
-func (s *ChannelConfigService) GetByAgent(agentName string) (*ChannelConfigDTO, error) {
-	agentName = strings.TrimSpace(agentName)
-	if agentName == "" {
-		return nil, nil
-	}
-	var r models.ChannelConfig
-	err := s.db.Where("agent_name = ?", agentName).First(&r).Error
 	if errors.Is(err, gorm.ErrRecordNotFound) {
 		return nil, nil
 	}
@@ -534,15 +504,6 @@ func (s *ChannelConfigService) syncProjectPmLeader(projectID, agentName string) 
 			"pm_leader_enabled": true,
 			"updated_at":        time.Now(),
 		}).Error
-}
-
-func (s *ChannelConfigService) projectHasPrimary(projectID string) (bool, error) {
-	var n int64
-	if err := s.db.Model(&models.ChannelConfig{}).
-		Where("project_id = ? AND is_primary = ?", projectID, true).Count(&n).Error; err != nil {
-		return false, err
-	}
-	return n > 0, nil
 }
 
 func (s *ChannelConfigService) defaultPrimaryAgent(projectID string) string {
